ADDED: Class based component UserSearch
Adds a class based React Component implementation version of the UserSearch to experiment with the benefits and drawbacks in practice.

In particular the ability to provide the data set to the generic is a powerful aspect as the typesafety aspect ensures that any data received from any endpoint will be provided in the appropriate structure.