CHANGED: visual output of package list
This adds a visual output of the package list as well as loading and error states to make this whole sandbox work from end to end.

Thanks to Redux, this also has automatic caching and even after-the-fact offline support making it work even if the network drops into offline mode.